requirements.txt
torch==2.2.1
transformers==4.46.2
sentencepiece==0.2.0
accelerate==0.26.0
numpy==1.26.0
protobuf==5.28.3
pyyaml==6.0.2